Refactor out more types into types package, and pull genesis block code into gen package
