+++ chain/state/statetree_test.go
@@ -1,4 +1,4 @@
-package chain
+package state
 
 import (
 	"testing"
