+++ chain/types/fullblock.go
@@ -1,0 +1,12 @@
+package types
+
+import "github.com/ipfs/go-cid"
+
+type FullBlock struct {
+	Header   *BlockHeader
+	Messages []*SignedMessage
+}
+
+func (fb *FullBlock) Cid() cid.Cid {
+	return fb.Header.Cid()
+}
